Tests: fix PHPUnit deprecation notices [1]
Removes the key from the entries returned by the `TypeProviderHelper` methods.

This puts the responsibility for having names for the individual entries in each dataset on the data provider methods in the tests (if so desired).
This prevents the `TypeProviderHelper` class having influence on what the parameter names for test methods should be.

This fixes a number of PHPUnit (10.5+) deprecation notices along the lines of "Providing invalid named argument $input for method WpOrg\Requests\Tests\*\*Test::test*() is deprecated and will not be supported in PHPUnit 11.0.".

Includes minor documentation fix for the affected methods in the `TypeProviderHelper` class.