Tests: fix PHPUnit deprecation notices [2]
Ensure the keys in datasets match the parameter names of the test methods to which the datasets are being passed.

This fixes a number of PHPUnit (10.5+) deprecation notices along the lines of "Providing invalid named argument $input for method WpOrg\Requests\Tests\*\*Test::test*() is deprecated and will not be supported in PHPUnit 11.0.".